ci: update aio payload size golden to reflect Angular update
Updates the AIO payload size golden to reflect the Angular
update as proposed by Renovate (see previous commit).

The main bundle got reduced and exceeded the threshold, so we needed
to update the golden. As part of this, all other bundle metrics are
updated even though not exceeding the threshold yet.